refactor: combine repos reducers
