+++ client/components/RepositoryBrowser.jsx
@@ -3,7 +3,7 @@ import classes from 'classnames'
 import { Row, Button, Col, Alert } from 'react-bootstrap'
 
 import RepositoryList from './RepositoryList.jsx'
-import RepositoryDetail from './RepositoryDetail.jsx'
+import RepositoryDetail from './../containers/RepositoryDetail.jsx'
 
 export default class RepositoryBrowser extends Component {
   static propTypes = {
@@ -13,22 +13,18 @@ export default class RepositoryBrowser extends Component {
     filterRepos: PropTypes.func.isRequired
   };
 
-  static defaultProps = {
-    repos: {items: {}, filterBy: ''}
-  };
-
   onFetchAll() {
     this.props.fetchAll(true)
   }
 
   render() {
     const {selected, repos, fetchAll, filterRepos} = this.props
-    const {isFetching} = repos
+    const {isFetching} = repos.status
     const selectedRepo = repos.items[selected]
     let content = null
 
     if (selected && selectedRepo) {
-      content = <RepositoryDetail repository={selectedRepo}/>
+      content = <RepositoryDetail checks={repos.checks} repository={selectedRepo}/>
     } else if (selected) {
       content = <Alert bsStyle='danger'>We didn’t find a repository {selected}.
         <Button
@@ -51,7 +47,7 @@ export default class RepositoryBrowser extends Component {
         <Col sm={4} className='col-sm-pull-8'>
           <RepositoryList isUpdating={isFetching}
                           repositories={repos.items}
-                          filterBy={repos.filterBy}
+                          filterBy={repos.status.filterBy}
                           filterRepos={filterRepos}
                           fetchAll={fetchAll}
                           selected={selected}/>
