+++ client/reducers/index.js
@@ -3,11 +3,9 @@ import { combineReducers } from 'redux'
 import auth from './auth'
 import user from './user'
 import repos from './repos'
-import checks from './checks'
 
 export default combineReducers({
   auth,
   user,
-  repos,
-  checks
+  repos
 })
