+++ client/reducers/repos/index.js
@@ -1,0 +1,11 @@
+import { combineReducers } from 'redux'
+
+import items from './items'
+import checks from './checks'
+import status from './status'
+
+export default combineReducers({
+  items,
+  checks,
+  status
+})
